Extract PostFragment
