+++ android/src/local/res/values/strings.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <resources>
     <string name="app_name">Happy News (Local)</string>
-    <string name="api_url">http://hsdabfjkhasdbjhsabdfjasdb/</string>
+    <string name="api_url">http://10.0.2.2:8080</string>
 </resources>
