+++ android/src/main/java/nl/fhict/happynews/android/activity/MainActivity.java
@@ -3,14 +3,8 @@ package nl.fhict.happynews.android.activity;
 import android.app.FragmentManager;
 import android.app.NotificationManager;
 import android.app.SearchManager;
-import android.content.Context;
-import android.content.DialogInterface;
 import android.content.Intent;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
 import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.AlertDialog;
 import android.support.v7.app.AppCompatActivity;
 import android.support.v7.widget.SearchView;
 import android.util.Log;
