+++ android/src/main/java/nl/fhict/happynews/android/fragments/MainFragment.java
@@ -1,150 +1,18 @@
 package nl.fhict.happynews.android.fragments;
 
-import android.app.Fragment;
-import android.content.Context;
-import android.content.DialogInterface;
-import android.net.ConnectivityManager;
-import android.net.NetworkInfo;
-import android.os.Bundle;
-import android.support.v4.widget.SwipeRefreshLayout;
-import android.support.v7.app.AlertDialog;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
-import android.view.LayoutInflater;
-import android.view.View;
-import android.view.ViewGroup;
-import nl.fhict.happynews.android.LoadListener;
-import nl.fhict.happynews.android.R;
-import nl.fhict.happynews.android.adapter.FeedAdapter;
+import com.koushikdutta.async.future.FutureCallback;
 import nl.fhict.happynews.android.manager.PostManager;
 import nl.fhict.happynews.android.model.Page;
-import nl.fhict.happynews.android.model.Post;
 
-import java.util.ArrayList;
-
-import static android.app.Activity.RESULT_OK;
-
-public class MainFragment extends Fragment implements LoadListener {
-
-    private PostManager postManager;
-    private SwipeRefreshLayout swipeRefresh;
-    private RecyclerView recyclerView;
-    private LinearLayoutManager layoutManager;
-    private FeedAdapter feedAdapter;
-
-    private boolean loading;
-    private int pastVisiblesItems;
-    private int visibleItemCount;
-    private int totalItemCount;
-
+public class MainFragment extends PostFragment {
 
     @Override
-    public View onCreateView(LayoutInflater inflater, ViewGroup container, Bundle savedInstanceState) {
-        View view = inflater.inflate(R.layout.fragment_main, container, false);
-
-        postManager = PostManager.getInstance(getActivity().getApplicationContext());
-
-        recyclerView = (RecyclerView) view.findViewById(R.id.recyclerView);
-        swipeRefresh = (SwipeRefreshLayout) view.findViewById(R.id.swipeRefresh);
-
-        feedAdapter = new FeedAdapter(getActivity(), new ArrayList<Post>());
-        layoutManager = new LinearLayoutManager(getActivity());
-        recyclerView.setAdapter(feedAdapter);
-        recyclerView.setLayoutManager(layoutManager);
-
-        postManager.setFeedAdapter(feedAdapter);
-
-        postManager.setFeedAdapter(feedAdapter);
-
-        // Display an error if not connected on start of this activity
-        checkConnection();
-
-        addScrollListener();
-
-        swipeRefresh.setOnRefreshListener(new SwipeRefreshLayout.OnRefreshListener() {
-            @Override
-            public void onRefresh() {
-                swipeRefresh.setRefreshing(false);
-                loading = false;
-                checkConnection();
-            }
-        });
-
-        return view;
-    }
-
-    private void addScrollListener() {
-        recyclerView.addOnScrollListener(new RecyclerView.OnScrollListener() {
-            @Override
-            public void onScrolled(RecyclerView recyclerView, int dx, int dy) {
-                if (dy > 0) { //Check if scrolled down
-                    visibleItemCount = layoutManager.getChildCount();
-                    totalItemCount = layoutManager.getItemCount();
-                    pastVisiblesItems = layoutManager.findFirstVisibleItemPosition();
-
-                    if (!loading) {
-                        if ((visibleItemCount + pastVisiblesItems) >= totalItemCount) {
-                            loading = true;
-                            Page lastPage = feedAdapter.getLastPage();
-                            if (!lastPage.isLast()) {
-                                postManager.load(
-                                    lastPage.getNumber() + 1,
-                                    PostManager.DEFAULT_PAGE_SIZE,
-                                    getActivity(),
-                                    MainFragment.this);
-                            }
-                        }
-                    }
-                }
-            }
-        });
-    }
-
-    /**
-     * When this activity is subscribed to a PostManager,
-     * this notification will be called when content
-     * is finished loading.
-     */
-    @Override
-    public void onFinishedLoading() {
-        loading = false;
-        swipeRefresh.setRefreshing(false);
+    protected void doRefresh(PostManager postManager, FutureCallback<Page> callback) {
+        postManager.refresh(getActivity(), callback);
     }
 
     @Override
-    public void onResume() {
-        super.onResume();
-        feedAdapter.notifyDataSetChanged();
-    }
-
-    /**
-     * Shows an error message when the app is not connected to the internet.
-     */
-    private void checkConnection() {
-        ConnectivityManager connectivityManager = (ConnectivityManager) getActivity().getSystemService(Context
-            .CONNECTIVITY_SERVICE);
-        NetworkInfo activeNetwork = connectivityManager.getActiveNetworkInfo();
-
-        boolean isConnected = activeNetwork != null && activeNetwork.isConnected();
-
-        if (!isConnected) {
-            AlertDialog alert = new AlertDialog.Builder(getActivity()).create();
-
-            alert.setTitle(getString(R.string.network_error_title));
-            alert.setMessage(getString(R.string.no_internet_message));
-            alert.setButton(RESULT_OK, getString(R.string.retry_button), new DialogInterface.OnClickListener() {
-                @Override
-                public void onClick(DialogInterface dialog, int which) {
-                    //Display an error if still not connected
-                    checkConnection();
-                }
-            });
-
-            alert.show();
-        } else {
-            swipeRefresh.setRefreshing(true);
-            loading = true;
-            postManager.refresh(getActivity(), MainFragment.this);
-        }
+    protected void doLoadNextPage(PostManager postManager, Page lastPage, FutureCallback<Page> callback) {
+        postManager.load(lastPage.getNumber() + 1, PostManager.DEFAULT_PAGE_SIZE, getActivity(), callback);
     }
 }
