+++ android/src/main/java/nl/fhict/happynews/android/manager/PostManager.java
@@ -1,7 +1,6 @@
 package nl.fhict.happynews.android.manager;
 
 import android.content.Context;
-import android.widget.Toast;
 import com.google.gson.Gson;
 import com.google.gson.GsonBuilder;
 import com.google.gson.JsonDeserializationContext;
@@ -11,9 +10,7 @@ import com.google.gson.JsonParseException;
 import com.google.gson.reflect.TypeToken;
 import com.koushikdutta.async.future.FutureCallback;
 import com.koushikdutta.ion.Ion;
-import nl.fhict.happynews.android.LoadListener;
 import nl.fhict.happynews.android.R;
-import nl.fhict.happynews.android.adapter.FeedAdapter;
 import nl.fhict.happynews.android.model.Page;
 
 import java.lang.reflect.Type;
@@ -44,11 +41,8 @@ public class PostManager {
     }
 
     private String apiUrl;
-    private FeedAdapter feedAdapter;
-    private Context context;
 
     private PostManager(Context context) {
-        this.context = context;
 
         apiUrl = context.getString(R.string.api_url);
 
@@ -68,19 +62,10 @@ public class PostManager {
      * @param page     The page number to load.
      * @param size     The amount of items on the page.
      * @param context  The context used to build the request.
-     * @param listener The listener to notify when loading is completed.
+     * @param callback The callback when the page is loaded.
      */
-    public void load(int page, int size, Context context, final LoadListener listener) {
-        loadPage(page, size, context, new FutureCallback<Page>() {
-            @Override
-            public void onCompleted(Exception e, Page result) {
-                feedAdapter.addPage(result);
-
-                if (listener != null) {
-                    listener.onFinishedLoading();
-                }
-            }
-        });
+    public void load(int page, int size, final Context context, final FutureCallback<Page> callback) {
+        loadPage(page, size, context, callback);
     }
 
     /**
@@ -90,19 +75,10 @@ public class PostManager {
      * @param page     The page number to load.
      * @param size     The amount of items on the page.
      * @param context  The context used to build the request.
-     * @param listener The listener to notify when loading is completed.
+     * @param callback The callback when the page is loaded.
      */
-    public void load(String query, int page, int size, Context context, final LoadListener listener) {
-        loadPage(query, page, size, context, new FutureCallback<Page>() {
-            @Override
-            public void onCompleted(Exception e, Page result) {
-                feedAdapter.addPage(result);
-
-                if (listener != null) {
-                    listener.onFinishedLoading();
-                }
-            }
-        });
+    public void load(String query, int page, int size, final Context context, final FutureCallback<Page> callback) {
+        loadPage(query, page, size, context, callback);
     }
 
 
@@ -110,21 +86,12 @@ public class PostManager {
      * Sends content of the first page to the FeedAdapter.
      *
      * @param context  The Application context.
-     * @param listener Implementing the LoadListener Interface.
+     * @param callback The callback when the page is loaded.
      */
-    public void refresh(final Context context, final LoadListener listener) {
+    public void refresh(final Context context, final FutureCallback<Page> callback) {
         final int firstPage = 0;
 
-        loadPage(firstPage, DEFAULT_PAGE_SIZE, context, new FutureCallback<Page>() {
-            @Override
-            public void onCompleted(Exception e, Page result) {
-                feedAdapter.setPage(result);
-
-                if (listener != null) {
-                    listener.onFinishedLoading();
-                }
-            }
-        });
+        loadPage(firstPage, DEFAULT_PAGE_SIZE, context, callback);
     }
 
     /**
@@ -132,25 +99,12 @@ public class PostManager {
      *
      * @param query The search query.
      * @param context  The Application context.
-     * @param listener Implementing the LoadListener Interface.
+     * @param callback The callback when the page is loaded.
      */
-    public void refresh(String query, final Context context, final LoadListener listener) {
+    public void refresh(String query, final Context context, final FutureCallback<Page> callback) {
         final int firstPage = 0;
 
-        loadPage(query, firstPage, DEFAULT_PAGE_SIZE, context, new FutureCallback<Page>() {
-            @Override
-            public void onCompleted(Exception e, Page result) {
-                if (e != null) {
-                    Toast.makeText(context, R.string.bad_request_toast, Toast.LENGTH_SHORT).show();
-                }
-
-                feedAdapter.setPage(result);
-
-                if (listener != null) {
-                    listener.onFinishedLoading();
-                }
-            }
-        });
+        loadPage(query, firstPage, DEFAULT_PAGE_SIZE, context, callback);
     }
 
 
@@ -201,15 +155,4 @@ public class PostManager {
             }
         });
     }
-
-    /**
-     * Assigns a feedAdapter to the PostManager.
-     *
-     * @param feedAdapter The feed adapter.
-     */
-    public void setFeedAdapter(FeedAdapter feedAdapter) {
-        this.feedAdapter = feedAdapter;
-    }
-
-
 }
