+++ android/src/main/res/layout/fragment_post.xml
